# platform-api-contract/src/main/resources/openapi/platform-api-v1.json
{
  "openapi" : "3.1.0",
  "info" : {
    "description" : "HTTP API of the multi-tenant application platform. Every successful body is {\"data\": ...} (collections add \"pagination\"); every failure is {\"error\": {code, message, fields?, requestId, traceId?}}.",
    "title" : "Platform API",
    "version" : "v1"
  },
  "servers" : [ {
    "url" : "/"
  } ],
  "paths" : {
...
    "/api/v1/auth/switch" : {
      "post" : {
        "description" : "Answers the destination host and a one-time proof, valid for a minute. The browser opens the destination's switch page with the proof after the #. An organization that does not exist and one the caller does not belong to give the same NOT_FOUND.",
        "operationId" : "switchOrganization",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SwitchOrganizationRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseSwitchTarget"
                }
              }
            },
            "description" : "OK"
          },
          "default" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "Error. The code says what went wrong; the message is safe to show."
          }
        },
        "summary" : "Ask to continue in another of your organizations",
        "tags" : [ "Authentication" ]
      }
    },
    "/api/v1/auth/switch/complete" : {
      "post" : {
        "description" : "Public, on an organization host: the proof from the previous host is the credential. On success answers 204 and sets the short-lived login cookie; the browser then continues with the sign-in navigation (/api/v1/auth/start). A proof that is unknown, used, expired, made for another organization, or whose person is not a member here is a VALIDATION_ERROR on the field token, always with the same message.",
        "operationId" : "completeSwitch",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CompleteSwitchRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK"
          },
          "default" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "Error. The code says what went wrong; the message is safe to show."
          }
        },
        "summary" : "Continue in this organization with the one-time proof",
        "tags" : [ "Authentication" ]
      }
    },
...
    "/api/v1/members" : {
      "get" : {
        "description" : "For the administrators of the organization of the host. NOT_FOUND on the platform host, FORBIDDEN for a member who is not an administrator.",
        "operationId" : "listMembers",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseListMemberView"
                }
              }
            },
            "description" : "OK"
          },
          "default" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "Error. The code says what went wrong; the message is safe to show."
          }
        },
        "summary" : "List the members of the organization",
        "tags" : [ "Members" ]
      }
    },
    "/api/v1/members/{membershipId}/administrator" : {
      "put" : {
        "description" : "A stop-gap until access policies exist. The last administrator cannot be released (CONFLICT). Only an active member can be an administrator.",
        "operationId" : "setMemberAdministrator",
        "parameters" : [ {
          "in" : "path",
          "name" : "membershipId",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SetAdministratorRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK"
          },
          "default" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "Error. The code says what went wrong; the message is safe to show."
          }
        },
        "summary" : "Name a member an administrator, or release them",
        "tags" : [ "Members" ]
      }
    },
    "/api/v1/members/{membershipId}/deactivate" : {
      "post" : {
        "description" : "The member is signed out of this organization at once and cannot get back in until reactivated; their other organizations are untouched. CONFLICT when already deactivated or when this is the last administrator. NOT_FOUND for a member of another organization.",
        "operationId" : "deactivateMember",
        "parameters" : [ {
          "in" : "path",
          "name" : "membershipId",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          },
          "default" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "Error. The code says what went wrong; the message is safe to show."
          }
        },
        "summary" : "Deactivate a member",
        "tags" : [ "Members" ]
      }
    },
    "/api/v1/members/{membershipId}/reactivate" : {
      "post" : {
        "description" : "The person can sign in again; sessions that ended do not come back and the administrator marker is not restored. CONFLICT when the member is not deactivated.",
        "operationId" : "reactivateMember",
        "parameters" : [ {
          "in" : "path",
          "name" : "membershipId",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          },
          "default" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "Error. The code says what went wrong; the message is safe to show."
          }
        },
        "summary" : "Reactivate a deactivated member",
        "tags" : [ "Members" ]
      }
    },
    "/api/v1/organizations" : {
      "get" : {
        "description" : "For the organization switcher, on any host: the answer is the caller's own memberships, not the organization the host names. Each entry carries the host the server built for it.",
        "operationId" : "listMyOrganizations",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseListOrganizationSummary"
                }
              }
            },
            "description" : "OK"
          },
          "default" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "Error. The code says what went wrong; the message is safe to show."
          }
        },
        "summary" : "The organizations the signed-in person is an active member of",
        "tags" : [ "Organizations" ]
      },
      "post" : {
        "description" : "The signed-in person creates an organization and becomes its founding administrator. The short name (slug) becomes the first label of the organization's host name. Answers 201 with the host to sign in at. A name or slug that is not acceptable or not free is a VALIDATION_ERROR on its field; a person who already founded as many organizations as allowed is FORBIDDEN. Only on the platform host.",
        "operationId" : "createOrganization",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateOrganizationRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseOrganizationCreated"
                }
              }
            },
            "description" : "OK"
          },
          "default" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "Error. The code says what went wrong; the message is safe to show."
          }
        },
        "summary" : "Found an organization",
        "tags" : [ "Organizations" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "ApiError" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "$ref" : "#/components/schemas/ErrorCode"
          },
          "fields" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "array",
              "items" : {
                "type" : "string"
              }
            }
          },
          "message" : {
            "type" : "string"
          },
          "requestId" : {
            "type" : "string"
          },
          "traceId" : {
            "type" : "string"
          }
        },
        "required" : [ "code", "message", "requestId" ]
      },
      "ApiErrorResponse" : {
        "type" : "object",
        "properties" : {
          "error" : {
            "$ref" : "#/components/schemas/ApiError"
          }
        },
        "required" : [ "error" ]
      },
...
      "ApiResponseListMemberView" : {
        "type" : "object",
        "properties" : {
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/MemberView"
            }
          }
        },
        "required" : [ "data" ]
      },
      "ApiResponseListOrganizationSummary" : {
        "type" : "object",
        "properties" : {
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/OrganizationSummary"
            }
          }
        },
        "required" : [ "data" ]
      },
      "ApiResponseOrganizationCreated" : {
        "type" : "object",
        "properties" : {
          "data" : {
            "$ref" : "#/components/schemas/OrganizationCreated"
          }
        },
        "required" : [ "data" ]
      },
...
      "ApiResponseSwitchTarget" : {
        "type" : "object",
        "properties" : {
          "data" : {
            "$ref" : "#/components/schemas/SwitchTarget"
          }
        },
        "required" : [ "data" ]
      },
...
      "CompleteSwitchRequest" : {
        "type" : "object",
        "properties" : {
          "token" : {
            "type" : "string",
            "maxLength" : 200,
            "minLength" : 0
          }
        },
        "required" : [ "token" ]
      },
      "CreateOrganizationRequest" : {
        "type" : "object",
        "properties" : {
          "displayName" : {
            "type" : "string",
            "maxLength" : 200,
            "minLength" : 0
          },
          "slug" : {
            "type" : "string",
            "maxLength" : 100,
            "minLength" : 0
          }
        },
        "required" : [ "displayName", "slug" ]
      },
...
      "ErrorCode" : {
        "type" : "string",
        "description" : "Stable machine-readable error code. Branch on this, never on the message.",
        "enum" : [ "VALIDATION_ERROR", "MALFORMED_REQUEST", "UNAUTHENTICATED", "FORBIDDEN", "NOT_FOUND", "METHOD_NOT_ALLOWED", "NOT_ACCEPTABLE", "CONFLICT", "CONCURRENT_MODIFICATION", "PAYLOAD_TOO_LARGE", "UNSUPPORTED_MEDIA_TYPE", "RATE_LIMITED", "INTERNAL_ERROR", "SERVICE_UNAVAILABLE", "TENANT_UNAVAILABLE" ]
      },
...
      "MemberView" : {
        "type" : "object",
        "properties" : {
          "administrator" : {
            "type" : "boolean"
          },
          "displayName" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "foundingAdministrator" : {
            "type" : "boolean"
          },
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "since" : {
            "type" : "string",
            "format" : "date-time"
          },
          "status" : {
            "type" : "string"
          },
          "you" : {
            "type" : "boolean"
          }
        },
        "required" : [ "displayName", "email", "id", "since", "status" ]
      },
      "OrganizationCreated" : {
        "type" : "object",
        "properties" : {
          "displayName" : {
            "type" : "string"
          },
          "host" : {
            "type" : "string"
          },
          "slug" : {
            "type" : "string"
          }
        },
        "required" : [ "displayName", "host", "slug" ]
      },
      "OrganizationSummary" : {
        "type" : "object",
        "properties" : {
          "administrator" : {
            "type" : "boolean"
          },
          "displayName" : {
            "type" : "string"
          },
          "host" : {
            "type" : "string"
          },
          "slug" : {
            "type" : "string"
          }
        },
        "required" : [ "displayName", "host", "slug" ]
      },
...
      "SetAdministratorRequest" : {
        "type" : "object",
        "properties" : {
          "administrator" : {
            "type" : "boolean"
          }
        },
        "required" : [ "administrator" ]
      },
...
      "SwitchOrganizationRequest" : {
        "type" : "object",
        "properties" : {
          "slug" : {
            "type" : "string",
            "maxLength" : 40,
            "minLength" : 0
          }
        },
        "required" : [ "slug" ]
      },
      "SwitchTarget" : {
        "type" : "object",
        "properties" : {
          "host" : {
            "type" : "string"
          },
          "token" : {
            "type" : "string"
          }
        },
        "required" : [ "host", "token" ]
      },
...
    }
  }
}